requirements.txt
requests~=2.22.0
pandas
python-telegram-bot
